fix: components moved to folders
