--- src/components/HistoryScreen.js
@@ -1,11 +1,0 @@
-import BottomBar from "./BottomBar";
-import TopBar from "./TopBar";
-
-export default function HistoryScreen() {
-    return (
-        <>
-            <TopBar />
-            <BottomBar />
-        </>
-    );
-}
